updated pricing

+++ keycloak-scripts/build-all.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-# Make sure we're running from /opt/keycloak/keycloak-scripts
-if [ "$(pwd)" != "/opt/keycloak/keycloak-scripts" ]; then
-    echo "Please run this script from /opt/keycloak/keycloak-scripts"
+# Make sure cwd folder name is keycloak-scripts
+if [ "$(basename "$PWD")" != "keycloak-scripts" ]; then
+    echo "Please run this script from the keycloak-scripts folder"
     exit 1
 fi
 
@@ -12,11 +12,14 @@ if [ "$EUID" -eq 0 ]; then
     exit 1
 fi
 
-# if custom-scripts.jar exists here or in /opt/keycloak/providers/, warn the user
-if [ -f custom-scripts.jar ] || [ -f /opt/keycloak/providers/custom-scripts.jar ]; then
-    echo "Warning: custom-scripts.jar will be overwritten"
+# if /opt/keycloak/providers/custom-scripts.jar exists, warn the user
+if [ -f /opt/keycloak/providers/custom-scripts.jar ]; then
+    echo "Warning: /opt/keycloak/providers/custom-scripts.jar will be overwritten"
     echo "Please save a copy if needed"
 fi
+
+# Remove existing jar if it exists
+rm -f custom-scripts.jar
 
 # Build the jar
 echo "Building 2 custom scripts"
@@ -24,6 +27,12 @@ jar cf custom-scripts.jar \
     META-INF/ \
     registration-api-validator.js \
     token-map-fullname.js
+
+# If the jar was not built, show an error and exit
+if [ ! -f custom-scripts.jar ]; then
+    echo "Error: custom-scripts.jar was not built"
+    exit 1
+fi
 
 # Copy the jar to the providers directory
 cp custom-scripts.jar /opt/keycloak/providers/
